requirements.txt
torch~=2.5.1
torchvision
Pillow~=11.0.0
clip~=1.0